move hooks into index.js
